improve code for settlement timings and growth
found that cannot have identical recruitment assignments because they
would be assigned to the same settle_timing.  So now the number of
parameters for timing will be based on the number of unique recruitment
settlement timings.  Need to create a warning on this.

ALSO  NOTE:  need to check on empirical wt-at-age and birthseasons.