bycatch_fleets
rename one of the fcast containers

consider changes for bycatch fleets:
Considered Changes for Bycatch Fleets

currently bycatch fleets are excluded from direct calculation of F from
input catch, and from the catch logL

bycatch fleets have selectivity and retention functions, so even though
they are considered to have unknown catch levels,
this does not mean that their calculated retained catch is zero.

I am considering the following changes:

1.  provide switch so that bycatch only fleet:
(a) can have retained and discarded catch calculated normally, or
(b) all their catch will be assigned to discard

2.  Because MSY and Yield per recruit are calculated in terms of dead
catch, they currently include catch from bycatch fleets.
provide switch so that:
(a) bycatch only fleets are treated normally in benchmark and forecast
(b) F for bycatch only fleets is kept constant in benchmark and
forecast, so is not included in any forecast cap&allocation calculations
the level of F for bycatch only fleets will then need to be set as a
constant, or calculated as mean from a range of specified years

ALSO:
provide for max_catch_area in forecast to be year-specific and to have
option of doing as a max or as a target.