Make excel file SS_330_Change_Log.xlsx the only change log, edit for clarity
It was confusing to have both a text (.tpl) and excel sheet as change logs, as
it was unclear which one was correct if they differed. To reconcile this and
reduce the amount of maintence required upon changes to SS, the .tpl is removed
so that changes to SS exes only need to be recorded in the excel document.

In addition, formatting changes are made to the excel document for clarity,
including: 1. Additional columns added to allow sorting based on other keywords,
2. sort so that the list is in reverse chronological order.