Add in text version of Change Log
Before this commit, both text and excel versions of change logs were removed
from master. This commit adds back in the change log which is intended to track
changes in the SS exe.

A text file was chosen so that it could be more easily managed than a
spreadsheet with version control.