Revised pull request template
Simplified the pull request template with redundancies from the issue template. Can add those sections back if requested. Major changes to note are: 1) The doc/r4ss/ssi section has been removed - issues will not automatically be created in those repos from a pull request or from an issue. It seemed like this feature wasn't working for everyone which is why it was removed, however I'm open to suggestions/putting it back in if people do like that feature. 2) The section to revise the change log if needed. It didn't look like this feature was being used plus with us moving to keeping most info in the issue, it seemed unnecessary, but again, if this feature is useful, I will gladly put it back in.

Another change to note: I also used block quotes for free text input sections, this was a stylistic choice as I thought it looked better than having plain text.

Please provide any comments/suggestions.