Revised pull request template (#434)
* Revised pull request template

Simplified the pull request template with redundancies from the issue template. 

* Attempts to emphasize using Issues vs PR

I tried to emphasize that users should put information in the issue rather than the pull request. And, changed "pull request" to PR in the comments and added "(s)" after issue to emphasize that multiple issues can be in a single PR.
Added extra return lines to the commented code block so it is more clear what should stay and what users delete to uncomment the code block.

---------

Co-authored-by: Kelli Johnson <kelli.johnson@noaa.gov>